+++ docs/changelog.md
@@ -31,3 +31,5 @@
 - Terminado a aula 02 - Componentização - Componente Summary
 - Terminado a aula 03 - Componentização - TransactionsTable
 - Terminado a aula 01 - Consumindo API - Criando frontend sem backend
+- Terminado a aula 02 - Consumindo API - Configurando o MirageJS
+- Terminado a aula 03 - Consumindo API - Configurando o Axios